refactoring by review: - combine labels structs - drop subscriptions/apps/channels bundle keys from types
